feat: reorganize component imports to use _components directory for better structure
