--- apps/web/src/app/(dashboard)/mod/products/manage-products/[id]/edit/page.tsx
@@ -1,3 +1,0 @@
-export default function EditProductPage() {
-  return <div>Hello</div>;
-}
